# Conflicts:
#	sql/base/characters.sql
#	sql/base/mangos.sql
#	src/game/CMakeLists.txt
#	src/game/Entities/Unit.cpp
#	src/game/Server/WorldSession.cpp
#	src/game/Server/WorldSocket.cpp
#	src/game/World/World.cpp
#	src/shared/Log.cpp
#	src/shared/revision_sql.h

+++ src/game/vmap/BIH.cpp
@@ -23,7 +23,7 @@
 void BIH::buildHierarchy(std::vector<uint32>& tempTree, buildData& dat, BuildStats& stats)
 {
     // create space for the first node
-    tempTree.push_back(3 << 30); // dummy leaf
+    tempTree.push_back(static_cast<uint32>(3 << 30)); // dummy leaf
     tempTree.insert(tempTree.end(), 2, 0);
     // tempTree.add(0);
